refactor: general refactor of the application;
refactored: enemyTurn, changeStatsModal, deck, field, scoreboard, graveyard, toolBar, miniField, atoms, hooks, more.

refs: n/a
