+++ src/shared/actions/mainPhaseActions.jsx
@@ -1,9 +1,9 @@
-export const mainPhaseMonster = [
-  "Enemy summons 1 monster",
+export const mainPhaseMonster = ["Enemy summons 1 monster"];
+
+export const mainPhaseMonsterWithTribute = [
+  "Enemy tribute summon a monster",
   "Enemy summons NO monsters",
 ];
-
-export const mainPhaseMonsterWithTribute = ["Enemy tribute summon a monster"];
 
 export const mainPhaseMagicTrap = [
   "Enemy set NO magic/trap cards",
